safari-600.7-branch/LayoutTests/fast/viewport/viewport-21-expected.txt

ALERT: viewport size 980x1078 scale 10 with limits [10, 10] and userScalable true
